BallotBlaze_Server/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
psycopg2-binary>=2.9,<3.0
channels>=4.0,<5.0
channels-redis>=4.0,<5.0
drf-yasg>=1.21,<2.0
asgiref>=3.6,<4.0